+++ themes/theme-rui/src/components/index.ts
@@ -10,6 +10,7 @@ export { DateField } from './DateField.styles';
 export { DatePicker } from './DatePicker.styles';
 export { Dialog } from './Dialog.styles';
 export { Divider } from './Divider.styles';
+export { Drawer } from './Drawer.styles';
 export { IconButton } from './IconButton.styles';
 export { Input } from './Input.styles';
 export { Field } from './Field.styles';